# service/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My Task List CRUD",
        "description": "My Task List Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8090",
    "basePath": "/api",
    "tags": [
        {
            "name": "Task List",
            "description": "API for tasks in the system"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all users in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/create": {
            "post": {
                "tags": [
                    "Tasks to add"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User that we want to create",
                        "schema": {
                            "required": [
                                "name",
                                "description"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "summary": "add new task in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/update": {
            "put": {
                "tags": [
                    "Tasks to update"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User that we want to update",
                        "schema": {
                            "required": [
                                "id",
                                "name",
                                "description"
                            ],
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "summary": "update task in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/delete": {
            "put": {
                "tags": [
                    "Tasks to delete"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User that we want to update",
                        "schema": {
                            "required": [
                                "id"
                            ],
                            "properties": {
                                "id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "summary": "delete task in system",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}